requirements.txt
python-telegram-bot==20.*
pydantic-settings==2.1.*

# subgram sdk
httpx
pydantic